mqtt_listener/requirements.txt
paho-mqtt==1.6.1
python-dotenv==0.21.0
influxdb-client==1.35.0